requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
upstash-redis>=1.2.0
